Fixed tests and cleaned up repo

+++ src/models/lorenz_96_model.py
@@ -9,7 +9,14 @@ from src.transformers import BoundedVarTransformer
 from typing import List
 
 class Lorenz96SDE(ProbabilisticModel, Continuous):
-    def __init__(self, name:str='Lorenz96SDE', seed:int=42, simulation_method='euler', ts=torch.linspace(0, 1.5, 21), priors:List=None, y0=torch.tensor([6.4558, 1.1054, -1.4502, -0.1985, 1.1905, 2.3887, 5.6689, 6.7284])):
+    def __init__(self, 
+                 name:str='Lorenz96SDE', 
+                 seed:int=42, 
+                 simulation_method='euler', 
+                 ts=torch.linspace(0, 1.5, 21), 
+                 priors:List=None, 
+                 y0=torch.tensor([6.4558, 1.1054, -1.4502, -0.1985, 1.1905, 2.3887, 5.6689, 6.7284],dtype=torch.float64)
+                 ):
         self.seed = seed
         self.param_dim = 3
         self.ts = ts
